refactor(ops): extract orders page pick-state, formatters, modals, print sheet into components (Phase 1, no behavior change)
- src/components/ops/orders/{types,format,shortage,usePickChecks}.ts
- ReviewRequestModal, ShortageListModal, print/PickSheetPrint extracted verbatim
- OrderRow/MobileOrderRow share one usePickChecks hook (same keys, same 400ms debounce)
- page.tsx: 2,484 → 1,783 lines
